# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Postos API",
        "description": "API de gestão de postos (PinPoint): JWT, RBAC, stock e histórico.\r\n\r\n**Rotas públicas (sem JWT):**\r\n- `GET /postos` — listar, paginar (`?page=`) ou detalhe (`?id=`)\r\n- `GET /postos/search` — pesquisa rápida (`?q=`, opcional `limit`, `tipo`, `combustivel`, `status`)\r\n\r\n**Papéis locais (`users.role_id`):** ver `docs/ROLES.md` e `GET /api/roles`.",
        "contact": {
            "email": "suporte@postosapi.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Prefixo das rotas API"
        }
    ],
    "paths": {
...
        "/posts/{post}/promotions": {
            "get": {
                "tags": [
                    "Gestor",
                    "Promoções"
                ],
                "summary": "Listar promoções locais do posto (RN-G-002)",
                "operationId": "4fbffb10894fdf753cd3551f3a31f21f",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "scheduled",
                                "active",
                                "ended",
                                "cancelled"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de promoções"
                    },
                    "403": {
                        "description": "Sem permissão"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Gestor",
                    "Promoções"
                ],
                "summary": "Criar promoção local (não-combustível)",
                "operationId": "f5eda117386cd947cf645afbde3d05aa",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePromotionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Promoção criada"
                    },
                    "422": {
                        "description": "Validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{post}/promotions/{promotion}": {
            "patch": {
                "tags": [
                    "Gestor",
                    "Promoções"
                ],
                "summary": "Actualizar promoção (antes de terminar)",
                "operationId": "921c240335efdc0a7bbadb48bb81d238",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "promotion",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePromotionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Promoção actualizada"
                    },
                    "404": {
                        "description": "Não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{post}/promotions/{promotion}/cancel": {
            "post": {
                "tags": [
                    "Gestor",
                    "Promoções"
                ],
                "summary": "Cancelar promoção",
                "operationId": "a0ca5d2454d7ab84438093d94ac199ab",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "promotion",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Promoção cancelada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{post}/stock-alerts": {
            "get": {
                "tags": [
                    "Gestor",
                    "Stock"
                ],
                "summary": "Alertas preditivos de rutura (RN-G-004)",
                "description": "Lista alertas activos/reconhecidos. Parâmetro `status` opcional. Use POST analyze para forçar análise imediata.",
                "operationId": "cdcd855068de30c19196e9902742d0f5",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "acknowledged",
                                "resolved"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alertas"
                    },
                    "403": {
                        "description": "Sem permissão"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{post}/stock-alerts/analyze": {
            "post": {
                "tags": [
                    "Gestor",
                    "Stock"
                ],
                "summary": "Executar análise preditiva agora",
                "operationId": "e03877df0b6346e034b68910ad20b5cf",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Análise concluída"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/posts/{post}/stock-alerts/{alert}/acknowledge": {
            "patch": {
                "tags": [
                    "Gestor",
                    "Stock"
                ],
                "summary": "Reconhecer alerta de rutura",
                "operationId": "a2e6dad1486fff0617d27fe07df2e4ef",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "alert",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alerta reconhecido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StorePromotionRequest": {
                "required": [
                    "title",
                    "discount_percent",
                    "starts_at",
                    "ends_at"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Lavagem -20%",
                        "maxLength": 120
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "discount_percent": {
                        "type": "number",
                        "format": "float",
                        "example": 15
                    },
                    "service_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "product_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "starts_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "ends_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "UpdatePromotionRequest": {
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "discount_percent": {
                        "type": "number",
                        "format": "float"
                    },
                    "starts_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "ends_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT no header Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}